Move `output_assert.rs` to `test_utils` submodule
