+++ crates/forge/tests/e2e/fork_warning.rs
@@ -4,7 +4,7 @@ use axum::{extract::Query, response::Redirect, routing::any, Router};
 use indoc::{formatdoc, indoc};
 use lazy_static::lazy_static;
 use shared::consts::EXPECTED_RPC_VERSION;
-use shared::output_assert::assert_stdout_contains;
+use shared::test_utils::output_assert::assert_stdout_contains;
 use std::{thread::sleep, time::Duration};
 use tokio::{
     net::TcpListener,
