+++ crates/forge/tests/e2e/forking.rs
@@ -3,7 +3,7 @@ use crate::e2e::common::runner::{
 };
 use forge::shared_cache::CACHE_DIR;
 use indoc::indoc;
-use shared::output_assert::assert_stdout_contains;
+use shared::test_utils::output_assert::assert_stdout_contains;
 
 #[test]
 fn without_cache() {
