+++ crates/forge/tests/e2e/steps.rs
@@ -1,6 +1,6 @@
 use crate::e2e::common::runner::{setup_package, test_runner};
 use indoc::indoc;
-use shared::output_assert::assert_stdout_contains;
+use shared::test_utils::output_assert::assert_stdout_contains;
 
 #[test]
 fn should_allow_less_than_default() {
