+++ crates/forge/tests/e2e/workspaces.rs
@@ -1,6 +1,6 @@
 use crate::e2e::common::runner::{setup_hello_workspace, setup_virtual_workspace, test_runner};
 use indoc::indoc;
-use shared::output_assert::assert_stdout_contains;
+use shared::test_utils::output_assert::assert_stdout_contains;
 use std::path::PathBuf;
 
 #[test]
