+++ crates/shared/src/lib.rs
@@ -9,8 +9,7 @@ pub mod consts;
 pub mod print;
 pub mod rpc;
 pub mod utils;
-
-pub mod output_assert;
+pub mod test_utils;
 
 pub async fn verify_and_warn_if_incompatible_rpc_version(
     client: &JsonRpcClient<HttpTransport>,
