+++ crates/shared/src/test_utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod output_assert;
